moves filters and paginator to questionnaires
